requirements.txt
anthropic>=0.40.0
appium-python-client>=3.1.0
pytest>=8.0.0
pytest-json-report>=1.5.0
selenium>=4.18.0
